+++ src/testData/mockData.json
@@ -315,10 +315,16 @@
       {
         "METHOD": "POST",
         "REQUEST": {
-          "id":1,
-          "jsonrpc":"2.0",
-          "method":"GetSmartContractSubState",
-          "params":["9611c53be6d1b32058b2747bdececed7e1216793","records",["0x1cc365ffd60bb50538e01d24c1f1e26c887c36f26a0de250660b8a1465c60667"]]
+          "id": 1,
+          "jsonrpc": "2.0",
+          "method": "GetSmartContractSubState",
+          "params": [
+            "9611c53be6d1b32058b2747bdececed7e1216793",
+            "records",
+            [
+              "0x1cc365ffd60bb50538e01d24c1f1e26c887c36f26a0de250660b8a1465c60667"
+            ]
+          ]
         },
         "RESPONSE": {
           "id": 1,
@@ -349,7 +355,7 @@
           "id": 1,
           "jsonrpc": "2.0",
           "result": {
-            "records":{
+            "records": {
               "crypto.ETH.address": "0xaa91734f90795e80751c96e682a321bb3c1a4186",
               "crypto.BTC.address": "1NZKHwpfqprxzcaijcjf71CZr27D8osagR"
             }
